chat evaluate response message 추가
