+++ src/app/(home)/mypage/page.tsx
@@ -7,6 +7,7 @@ import { Session } from "next-auth";
 import { Spin } from "antd";
 
 import useAuth from "@/src/lib/auth";
+import useInfo from "@/src/utils/handleGetInfo";
 import { LoadingOutlined } from "@ant-design/icons";
 import Mypage from "@/src/containers/mypage";
 import { userInfoState } from "@/src/lib/recoil";
@@ -15,26 +16,11 @@ import { UserInfo } from "@/src/types/type";
 export default function Home() {
   const router = useRouter();
   const { authFetch } = useAuth();
+  const { getInfo } = useInfo();
 
   const [recoilUserInfoState, setUserInfo] = useRecoilState(userInfoState);
 
   const [userInfo, setRecoilUserInfoState] = useState<Session | null>(null);
-
-  const getInfo = async () => {
-    try {
-      const res = await authFetch("/api/user/info", {
-        method: "GET",
-      });
-      const data: UserInfo = await res.json();
-      setUserInfo((prevState) => {
-        if (!prevState) return prevState;
-        return {
-          ...prevState,
-          user: data,
-        };
-      });
-    } catch (error) {}
-  };
 
   useEffect(() => {
     setRecoilUserInfoState(recoilUserInfoState);
